Move `mpu401.cpp|h` to `src/hardware/audio`
