+++ src/hardware/CMakeLists.txt
@@ -7,6 +7,7 @@ target_sources(libdosboxcommon PRIVATE
   audio/imfc.cpp
   audio/innovation.cpp
   audio/lpt_dac.cpp
+  audio/mpu401.cpp
   audio/opl.cpp
   audio/pcspeaker.cpp
   audio/pcspeaker_discrete.cpp
@@ -70,7 +71,6 @@ target_sources(libdosboxcommon PRIVATE
   ipx.cpp
   ipxserver.cpp
   memory.cpp
-  mpu401.cpp
   ne2000.cpp
   pci_bus.cpp
   pic.cpp
