+++ src/hardware/meson.build
@@ -7,6 +7,7 @@ libhardware_sources = files(
     'audio/imfc.cpp',
     'audio/innovation.cpp',
     'audio/lpt_dac.cpp',
+    'audio/mpu401.cpp',
     'audio/opl.cpp',
     'audio/pcspeaker.cpp',
     'audio/pcspeaker_discrete.cpp',
@@ -71,7 +72,6 @@ libhardware_sources = files(
     'ipx.cpp',
     'ipxserver.cpp',
     'memory.cpp',
-    'mpu401.cpp',
     'ne2000.cpp',
     'pci_bus.cpp',
     'pic.cpp',
